@@ -16,31 +16,31 @@
     "checks": "npm run lint && npm run test"
   },
   "devDependencies": {
-    "autoprefixer": "^6.4.1",
-    "babel-core": "^6.14.0",
-    "babel-preset-es2015": "^6.14.0",
-    "babel-preset-react": "^6.11.1",
-    "babel-register": "^6.14.0",
+    "autoprefixer": "^6.5.0",
+    "babel-core": "^6.17.0",
+    "babel-preset-es2015": "^6.16.0",
+    "babel-preset-react": "^6.16.0",
+    "babel-register": "^6.16.3",
     "babelify": "^7.3.0",
     "browserify": "^13.1.0",
     "envify": "^3.4.1",
     "enzyme": "^2.4.1",
-    "eslint": "^3.4.0",
-    "eslint-config-bitcrowd-react": "^10.0.1",
-    "eslint-plugin-import": "^1.14.0",
-    "eslint-plugin-jsx-a11y": "^2.2.1",
-    "eslint-plugin-react": "^6.2.0",
+    "eslint": "^3.7.1",
+    "eslint-config-bitcrowd-react": "^12.0.0",
+    "eslint-plugin-import": "^2.0.1",
+    "eslint-plugin-jsx-a11y": "^2.2.2",
+    "eslint-plugin-react": "^6.3.0",
     "gulp": "^3.9.1",
-    "gulp-postcss": "^6.1.1",
+    "gulp-postcss": "^6.2.0",
     "gulp-sass": "^2.3.2",
-    "gulp-sourcemaps": "^1.6.0",
+    "gulp-sourcemaps": "^2.0.0",
     "gulp-uglify": "^2.0.0",
     "gulp-zip": "^3.2.0",
-    "jasmine": "=2.4.1",
-    "jsdom": "^9.5.0",
-    "nodemon": "^1.10.2",
+    "jasmine": "^2.5.2",
+    "jsdom": "^9.6.0",
+    "nodemon": "^1.11.0",
     "nyc": "^8.3.1",
-    "react-addons-test-utils": "^15.3.1",
+    "react-addons-test-utils": "^15.3.2",
     "rimraf": "^2.5.4",
     "through2": "^2.0.1",
     "vinyl-buffer": "^1.0.0",
@@ -48,11 +48,11 @@
   },
   "dependencies": {
     "bootstrap": "^4.0.0-alpha.4",
-    "jquery": "^3.1.0",
+    "jquery": "^3.1.1",
     "octicons": "^4.4.0",
-    "react": "^15.3.1",
-    "react-dom": "^15.3.1",
-    "react-router": "^2.7.0",
-    "speakingurl": "^9.0.0"
+    "react": "^15.3.2",
+    "react-dom": "^15.3.2",
+    "react-router": "^2.8.1",
+    "speakingurl": "^10.0.0"
   }
 }
