@@ -10,7 +10,7 @@ describe('ticket-list', () => {
 
   it('renders a list of tickets', () => {
     const wrapper = shallow(<TicketList tickets={tickets} />);
-    const items = tickets.map((t) => (<TicketListItem ticket={t} />));
+    const items = tickets.map(t => (<TicketListItem ticket={t} />));
     expect(wrapper.find('ul').containsAllMatchingElements(items)).toBe(true);
   });
 });
@@ -34,16 +34,16 @@ describe('ticket-list-item', () => {
 
   it('renders a copy-button for the commit messsage', () => {
     const buttons = wrapper.find(CopyButton);
-    expect(buttons.someWhere((b) => b.prop('value') === fmt.commit(ticket))).toBe(true);
+    expect(buttons.someWhere(b => b.prop('value') === fmt.commit(ticket))).toBe(true);
   });
 
   it('renders a copy-button for the branch name', () => {
     const buttons = wrapper.find(CopyButton);
-    expect(buttons.someWhere((b) => b.prop('value') === fmt.branch(ticket))).toBe(true);
+    expect(buttons.someWhere(b => b.prop('value') === fmt.branch(ticket))).toBe(true);
   });
 
   it('renders a copy-button for the git commands', () => {
     const buttons = wrapper.find(CopyButton);
-    expect(buttons.someWhere((b) => b.prop('value') === fmt.command(ticket))).toBe(true);
+    expect(buttons.someWhere(b => b.prop('value') === fmt.command(ticket))).toBe(true);
   });
 });
