@@ -11,15 +11,18 @@ function search(adapters, loc, doc, fn) {
   const results = [];
 
   function done() {
-    const res = results.find((e) => (e !== null));
+    const res = results.find(e => (e !== null));
     fn(res || null);
   }
 
   adapters.forEach((adapter, i) => {
     adapter.inspect(loc, doc, (err, res) => {
       if (err) console.error(err); // eslint-disable-line no-console
+
       results[i] = err ? null : res;
-      if (--pending === 0) done();
+      pending -= 1;
+
+      if (pending === 0) done();
     });
   });
 }
